♻️ Decouple collections in inverntory plugins
This puts upstream and downstream collection references into separate
inventory plugins instead of keeping both mappings in some and decide
what to use at runtime. Previously, that decision was based on what
the license discovery returned and it does not belong here. Now, the
collection to be used is governed by what plugin is `pip install`ed.

As a side effect, this also drops the use of
`detect_server_product_name()` from
`awx_plugins.interfaces._temporary_private_licensing_api`, so we'll
be able to drop it from `awx_plugins.interfaces` and put back into
awx.